refactor(fe): improve imports, use props and composables
