+++ src/modules/faqs/components/index.ts
@@ -1,0 +1,2 @@
+export { default as Tiptap } from './tiptap.vue';
+export { default as DataAction } from './data-action.vue';
