+++ src/modules/faqs/index.ts
@@ -1,3 +1,2 @@
-export { default as Tiptap } from './components/tiptap.vue';
-export { default as DataAction } from './components/data-action.vue';
-export * from './services/api';
+export * from './types';
+export * from './columns';
